Bower: Extract logic for 'stashing' directories into a utility class
Move this logic into an utility class in order to enable code re-use.
In order to make the client code more elegant it has been refacored to
a 'Closable' which eliminates the try and finally block.
Most importantly the Map variable holding the currently stashed
directories is not moved to the static context which is very beneficial
in terms of robustness.
